fix(dto): corrigindo o tipo do campo quantidade Integer -> Long para evitar erro de conversão

+++ src/main/java/com/academico/DataShape/model/dto/responses/findQtdTotalVendasResponse.java
@@ -1,5 +1,5 @@
 package com.academico.DataShape.model.dto.responses;
 import java.util.List;
-public record findQtdTotalVendasResponse(Integer total_un_vendidas, List<QtdVendaDTO> qtd_vendas_por_obra) {
+public record findQtdTotalVendasResponse(Long total_un_vendidas, List<QtdVendaDTO> qtd_vendas_por_obra) {
     
 }
